requirements.txt
streamlit
numpy
pillow
scikit-learn
opencv-python
tensorflow
pandas
matplotlib
